Move sphinx requirements to main requirements.txt
The goal is to make readthedocs read the correct list of requirements.
The readthedocs system can only read from one requirements.txt file, so
move the sphinx pre-requisites to the main requirements.txt file and
drop the one in the docs/ folder. Also, add a .readthedocs.yaml
configuration file to specify the requirements file.