Support analysis with lossy antenna-to-LNA feed
The losses between the antenna and the LNA can be neglected when
considering an LNBF (with an integrated feed horn). However, when the
feed is separate from the LNA (when using discrete components), the feed
loss can be significant.

This patch adds support for including such losses on the link budget
computation. It also refactors the calculations involving a generic
passive two-port attenuator, such as a waveguide, transmission line
(e.g., coax cable), and even the atmosphere when modeling the antenna
noise. More specifically, this patch adds dedicated functions to compute
the equivalent input/output noise temperature and noise figure of a
passive attenuator. Then, it reuses those functions where applicable.