Add sat latitude/altitude and Rx height options
Previously, the satellite was always assumed to be in geosynchronous
equatorial orbit with zero latitude and altitude 35,786 km. While this
continues to be the default, this patch allows for changing the
satellite latitude and altitude via new CLI parameters.

This patch also introduces a CLI parameter to define the Rx station's
height above mean sea level, which was previously fixed at zero.