Support satellite position prediction via TLE info
Use the Skyfield package's implementation to parse a TLE dataset and
predict the satellite position at a chosen epoch.

Add options to control the saving of TLE datasets and the search on
CelesTrak's database. Include also an option for defining the satellite
observation time to be used in the prediction.